+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Upcoming
+## Alpha 4.1.13
 * Increased damage taken by wither from blue skulls
   * Damage from blue skulls is multiplied by 0.8% -> 1% of its max health
 * Renamed default wither stats names in data pack
@@ -8,7 +8,7 @@
 * Charge Object
   * Added `second_phase` in Wither's charge object. If true, the wither will charge 3 times in rapid succession
   * Added `second_phase_times`, `second_phase_tick_reduction` and `second_phase_max_reduction` to configure the second phase charging times, tick reduction and max reduction
-  * Renamed `chance_on_hit` to `chance_on_hit`
+  * Renamed `max_chance` to `chance_on_hit`
   * Renamed `time` to `base_time`
 
 ## Alpha 4.1.12